+++ CopilotKit/packages/react-ui/src/components/chat/Chat.tsx
@@ -60,7 +60,6 @@ import { RenderAgentStateMessage as DefaultRenderAgentStateMessage } from "./mes
 import { RenderImageMessage as DefaultRenderImageMessage } from "./messages/RenderImageMessage";
 import { AssistantMessage as DefaultAssistantMessage } from "./messages/AssistantMessage";
 import { UserMessage as DefaultUserMessage } from "./messages/UserMessage";
-import { Suggestion } from "./Suggestion";
 import React, { useEffect, useRef, useState } from "react";
 import {
   SystemMessageFunction,
@@ -74,14 +73,17 @@ import { Message, Role, TextMessage, ImageMessage } from "@copilotkit/runtime-cl
 import { randomId } from "@copilotkit/shared";
 import {
   AssistantMessageProps,
+  ComponentsMap,
   InputProps,
   MessagesProps,
   RenderMessageProps,
+  RenderSuggestionsListProps,
   UserMessageProps,
 } from "./props";
 
 import { HintFunction, runAgent, stopAgent } from "@copilotkit/react-core";
 import { ImageUploadQueue } from "./ImageUploadQueue";
+import { Suggestions as DefaultRenderSuggestionsList } from "./Suggestions";
 
 /**
  * Props for CopilotChat component.
@@ -136,6 +138,12 @@ export interface CopilotChatProps {
    * A callback function for thumbs down feedback
    */
   onThumbsDown?: (message: string) => void;
+
+  /**
+   * A list of markdown components to render in assistant message.
+   * Useful when you want to render custom elements in the message (e.g a reference tag element)
+   */
+  markdownTagRenderers?: ComponentsMap;
 
   /**
    * Icons can be used to set custom icons for the chat window.
@@ -205,6 +213,11 @@ export interface CopilotChatProps {
    * A custom RenderImageMessage component to use instead of the default.
    */
   RenderImageMessage?: React.ComponentType<RenderMessageProps>;
+
+  /**
+   * A custom suggestions list component to use instead of the default.
+   */
+  RenderSuggestionsList?: React.ComponentType<RenderSuggestionsListProps>;
 
   /**
    * A custom Input component to use instead of the default.
@@ -291,12 +304,14 @@ export function CopilotChat({
   onCopy,
   onThumbsUp,
   onThumbsDown,
+  markdownTagRenderers,
   Messages = DefaultMessages,
   RenderTextMessage = DefaultRenderTextMessage,
   RenderActionExecutionMessage = DefaultRenderActionExecutionMessage,
   RenderAgentStateMessage = DefaultRenderAgentStateMessage,
   RenderResultMessage = DefaultRenderResultMessage,
   RenderImageMessage = DefaultRenderImageMessage,
+  RenderSuggestionsList = DefaultRenderSuggestionsList,
   Input = DefaultInput,
   className,
   icons,
@@ -482,20 +497,13 @@ export function CopilotChat({
         onCopy={handleCopy}
         onThumbsUp={onThumbsUp}
         onThumbsDown={onThumbsDown}
+        markdownTagRenderers={markdownTagRenderers}
       >
         {currentSuggestions.length > 0 && (
-          <div className="suggestions">
-            {currentSuggestions.map((suggestion, index) => (
-              <Suggestion
-                key={index}
-                title={suggestion.title}
-                message={suggestion.message}
-                partial={suggestion.partial}
-                className={suggestion.className}
-                onClick={(message) => handleSendMessage(message)}
-              />
-            ))}
-          </div>
+          <RenderSuggestionsList
+            onSuggestionClick={handleSendMessage}
+            suggestions={currentSuggestions}
+          />
         )}
       </Messages>