+++ CopilotKit/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,38 @@
 # ui
 
+## 1.8.13-next.3
+
+### Patch Changes
+
+- @copilotkit/react-core@1.8.13-next.3
+- @copilotkit/runtime-client-gql@1.8.13-next.3
+- @copilotkit/shared@1.8.13-next.3
+
+## 1.8.13-next.2
+
+### Patch Changes
+
+- @copilotkit/react-core@1.8.13-next.2
+- @copilotkit/runtime-client-gql@1.8.13-next.2
+- @copilotkit/shared@1.8.13-next.2
+
+## 1.8.13-next.1
+
+### Patch Changes
+
+- Updated dependencies [7fcf5c4]
+  - @copilotkit/react-core@1.8.13-next.1
+  - @copilotkit/runtime-client-gql@1.8.13-next.1
+  - @copilotkit/shared@1.8.13-next.1
+
+## 1.8.13-next.0
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.13-next.0
+- @copilotkit/react-core@1.8.13-next.0
+- @copilotkit/shared@1.8.13-next.0
+
 ## 1.8.12
 
 ### Patch Changes