+++ CopilotKit/packages/runtime/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @copilotkit/runtime
 
+## 1.8.13-next.3
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.13-next.3
+
+## 1.8.13-next.2
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.13-next.2
+
+## 1.8.13-next.1
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.13-next.1
+
+## 1.8.13-next.0
+
+### Patch Changes
+
+- 6ed54f4: - AG-UI 0.0.28
+  - @copilotkit/shared@1.8.13-next.0
+
 ## 1.8.12
 
 ### Patch Changes